requirements.txt
aioredis==1.2.0
appnope==0.1.0
asgi-redis==1.4.3
asgiref==3.1.1
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0rc1
Django==2.2
django-cors-headers==2.5.3
django-extensions==2.1.6
django-guardian==1.5.0
django-memoize==2.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
ecdsa==0.13
entrypoints==0.3
future==0.17.1
graphviz==0.10.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.1
msgpack-python==0.5.6
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
oauthlib==3.0.1
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pydotplus==2.0.2
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.8
pylint-plugin-utils==0.5
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.1
pyzmq==18.0.1
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
zope.interface==4.6.0
